+++ src/api_subject_space/requirements.txt
@@ -1,0 +1,12 @@
+fastapi[standard]==0.112.0
+uvicorn==0.30.6
+pydantic==2.8.2
+azure-servicebus==7.12.2
+python-dotenv==1.0.1
+azure-cosmos==4.7.0
+langchain-text-splitters==0.3.5
+langchain-openai==0.3.0
+langchain-community==0.3.14
+langchain-core==0.3.29
+azure-search-documents==11.5.2
+azure-identity==1.19.0
