+++ src/autopodcaster_model/autopodcaster_model/SubjectSpace.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+class SubjectSpace(BaseModel):
+    id: str
+    subject: str
+    date: str
+    last_updated: str
+    input_ids: list
+    index_name: str
