+++ src/autopodcaster_model/autopodcaster_model/__init__.py
@@ -1,3 +1,4 @@
 # __init__.py
 
 from .Input import Input
+from .SubjectSpace import SubjectSpace
