+++ src/autopodcaster_model/requirements.txt
@@ -1,2 +1,3 @@
 wheel==0.45.1
 setuptools==75.8.2
+pydantic==2.8.2
