--- src/subject_space/requirements.txt
@@ -1,10 +1,0 @@
-azure-servicebus==7.12.2
-fastapi==0.112.0
-python-dotenv==1.0.1
-azure-cosmos==4.7.0
-langchain_core==0.2.29
-langchain_text_splitters==0.2.2
-langchain_openai==0.1.21
-langchain_community==0.2.11
-azure-search-documents==11.5.1
-azure-identity==1.17.1
