+++ src/ui/AutoPodcaster/Services/IInputService.cs
@@ -1,4 +1,3 @@
-using System.Runtime.InteropServices;
 using AutoPodcaster.Models;
 
 public interface IInputService
