+++ src/ui/AutoPodcaster/Services/ISubjectSpaceService.cs
@@ -1,0 +1,9 @@
+using AutoPodcaster.Models;
+
+public interface ISubjectSpaceService
+{
+    Task<IQueryable<SubjectSpace>> GetSubjectSpacesAsync();
+    Task<SubjectSpace> GetSubjectSpaceByIdAsync(string id);
+    Task CreateSubjectSpaceAsync(string subject);
+    Task DeleteSubjectSpaceAsync(string id);
+}
